Move Gopay Notification to base class
The notification is moved so other Gopay implementations
can use the same notification processing.
This is handy for example for internal Gopay trial
implementation done in one of our integrations.

API handler for processing notifications is now
determining gateway based on the provided payment
identifier. This is necessary if someone uses own
Gopay implementation (e.g. trial) and needs the
notification to be processed through the trial gateway.

remp2020/crm-gopay-module#9